+++ builds/8chanSS.meta.js
@@ -1,6 +1,6 @@
 // ==UserScript==
 // @name         8chanSS
-// @version      1.49.0
+// @version      1.50.0
 // @namespace    8chanss
 // @description  A userscript to add functionality to 8chan.
 // @author       otakudude